+++ src/main/java/org/example/calculator/core/parser/nodes/NumberASTNode.java
@@ -1,18 +1,20 @@
 package org.example.calculator.core.parser.nodes;
 
+import java.math.BigDecimal;
+
 /**
  * 纯数字节点，也可以称为叶子节点
  */
 public class NumberASTNode implements ASTNode {
 
-    private final double value;
+    private final BigDecimal value;
 
-    public NumberASTNode(double value) {
+    public NumberASTNode(BigDecimal value) {
         this.value = value;
     }
 
     @Override
-    public double evaluate() {
+    public BigDecimal evaluate() {
         return value;
     }
 }
